Seperated Registry to own repository
